requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.0
ultralytics==8.0.220
torch==2.1.0
torchvision==0.16.0
insightface==0.7.3
onnxruntime==1.16.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
